requirements.txt
langchain
langchain-community
langchain-text-splitters    
flask
chromadb
pdfplumber
tiktoken
unstructured
pypdf
python-dotenv
numpy
pandas
ipykernel
pinecone
streamlit
deep-translator
SpeechRecognition
PyAudio
gTTS 
audio_recorder_streamlit
pydub
twilio
ragas

# pip install -U langchain-community